@@ -1,62 +1,62 @@
 {
-	"name": "profile",
-	"version": "1.0.0",
-	"scripts": {
-		"ng": "ng",
-		"start": "ng serve --open --port 6200",
-		"build": "ng build",
-		"watch": "ng build --watch --configuration development",
-		"test": "ng test",
-    "deploy:ghdocs": "ng build --prod --output-path docs --base-href /curriculum/"
-	},
-	"private": true,
-	"dependencies": {
-		"@angular/animations": "~13.0.0",
-		"@angular/cdk": "~13.0.0",
-		"@angular/common": "~13.0.0",
-		"@angular/compiler": "~13.0.0",
-		"@angular/core": "~13.0.0",
-		"@angular/forms": "~13.0.0",
-		"@angular/platform-browser": "~13.0.0",
-		"@angular/platform-browser-dynamic": "~13.0.0",
-		"@angular/router": "~13.0.0",
-		"@ngneat/transloco": "^3.1.3",
-		"@ngrx/router-store": "^13.0.2",
-		"@ngrx/store": "^13.0.2",
-		"@ngrx/store-devtools": "^13.0.2",
-		"chart.js": "^3.3.2",
-		"file-saver": "^2.0.5",
-		"jwt-decode": "^3.1.2",
-		"primeflex": "^3.1.0",
-		"primeicons": "5.0.0",
-		"primeng": "^13.0.4",
-		"prismjs": "1.9.0",
-		"rxjs": "~7.4.0",
-		"tslib": "^2.3.0",
-		"web-animations-js": "^2.3.2",
-		"xlsx": "^0.17.5",
-		"zone.js": "~0.11.4"
-	},
-	"devDependencies": {
-		"@angular-devkit/build-angular": "~13.0.4",
-		"@angular/cli": "~13.0.4",
-		"@angular/compiler-cli": "~13.0.0",
-		"@types/file-saver": "^2.0.5",
-		"@types/jasmine": "~3.10.0",
-		"@types/jasminewd2": "~2.0.8",
-		"@types/node": "^12.11.1",
-		"codelyzer": "^6.0.0",
-		"jasmine-core": "~3.10.0",
-		"jasmine-spec-reporter": "~5.0.0",
-		"karma": "~6.3.0",
-		"karma-chrome-launcher": "~3.1.0",
-		"karma-coverage": "~2.0.3",
-		"karma-coverage-istanbul-reporter": "~3.0.3",
-		"karma-jasmine": "~4.0.0",
-		"karma-jasmine-html-reporter": "~1.7.0",
-		"protractor": "~7.0.0",
-		"ts-node": "~8.3.0",
-		"tslint": "~6.1.0",
-		"typescript": "~4.4.3"
-	}
+  "name": "profile",
+  "author": "Lucas Peixoto",
+  "email": "lspeixotodev@gmail.com",
+  "license": "MIT",
+  "version": "2.0.0",
+  "private": true,
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve --open --port 6200",
+    "transloco:optimize": "transloco-optimize dist/profile/assets/i18n",
+    "build:prod": "ng build && npm run transloco:optimize",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test",
+    "ghdeploy": "ng build --output-path docs --base-href /profile/"
+  },
+  "dependencies": {
+    "@angular/animations": "^14.2.5",
+    "@angular/cdk": "^14.2.4",
+    "@angular/common": "^14.2.5",
+    "@angular/compiler": "^14.2.5",
+    "@angular/core": "^14.2.5",
+    "@angular/forms": "^14.2.5",
+    "@angular/platform-browser": "^14.2.5",
+    "@angular/platform-browser-dynamic": "^14.2.5",
+    "@angular/router": "^14.2.5",
+    "@ngrx/router-store": "^14.3.2",
+    "@ngrx/store": "^14.3.2",
+    "@ngrx/store-devtools": "^14.3.2",
+    "chart.js": "^3.3.2",
+    "primeflex": "^3.2.1",
+    "primeicons": "^6.0.1",
+    "primeng": "^14.1.2",
+    "prismjs": "1.9.0",
+    "rxjs": "^7.5.7",
+    "tslib": "^2.3.0",
+    "web-animations-js": "^2.3.2",
+    "zone.js": "~0.11.4"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^14.2.5",
+    "@angular/cli": "^14.2.5",
+    "@angular/compiler-cli": "^14.2.5",
+    "@types/jasmine": "~3.10.0",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.2",
+    "jasmine-core": "~3.10.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~4.8.4"
+  }
 }
