@@ -1,9 +1,14 @@
 import { RouterModule } from '@angular/router';
 import { NgModule } from '@angular/core';
 
-import { PersonalComponent } from './personal/personal/personal.component';
+import { PersonalComponent } from './profile/topics/personal/personal.component';
 import { AppMainComponent } from './app.main.component';
-import { DetailComponent } from './personal/detail/detail.component';
+import { AboutMeComponent } from './profile/topics/about-me/about-me.component';
+import { ExperiencesComponent } from './profile/topics/experiences/experiences.component';
+import { FormationsComponent } from './profile/topics/formations/formations.component';
+import { SkillsComponent } from './profile/topics/skills/skills.component';
+import { CertificationsComponent } from './profile/topics/certifications/certifications.component';
+import { ProjectsComponent } from './profile/topics/projects/projects.component';
 
 @NgModule({
 	imports: [
@@ -17,11 +22,30 @@ import { DetailComponent } from './personal/detail/detail.component';
 							path: '',
 							component: PersonalComponent,
 						},
-            {
-							path: 'cv/:item',
-							component: DetailComponent,
+						{
+							path: 'cv/sobre-min',
+							component: AboutMeComponent,
 						},
-
+						{
+							path: 'cv/experiencias',
+							component: ExperiencesComponent,
+						},
+						{
+							path: 'cv/certificacoes',
+							component: CertificationsComponent,
+						},
+						{
+							path: 'cv/formacoes',
+							component: FormationsComponent,
+						},
+						{
+							path: 'cv/competencias',
+							component: SkillsComponent,
+						},
+						{
+							path: 'cv/projetos',
+							component: ProjectsComponent,
+						},
 					],
 				},
 
