@@ -19,14 +19,13 @@ import { AppMenuComponent } from './app.menu.component';
 import { AppMenuitemComponent } from './app.menuitem.component';
 
 import { PrimeModule } from '@shared/modules/prime.module';
-import { TranslocoRootModule } from '@shared/modules/transloco-root.module';
+
 import { SharedCommonModule } from '@shared/modules/shared-common.module';
 
-import { CommonServices } from '@app/shared/services/common.service';
+import { CommonServices } from '@shared/services/common.service';
 import { ConfigService } from '@config/services/app.config.service';
 import { MenuService } from '@config/services/app.menu.service';
 import { environment } from '@envs/environment';
-
 
 @NgModule({
 	imports: [
@@ -36,10 +35,9 @@ import { environment } from '@envs/environment';
 		AppRoutingModule,
 		HttpClientModule,
 		BrowserAnimationsModule,
-		TranslocoRootModule,
 		SharedCommonModule,
 		PrimeModule,
-    StoreModule.forRoot(reducers),
+		StoreModule.forRoot(reducers),
 		StoreDevtoolsModule.instrument({
 			maxAge: 25,
 			logOnly: environment.production,
@@ -63,7 +61,6 @@ import { environment } from '@envs/environment';
 		MenuService,
 
 		{ provide: LOCALE_ID, useValue: 'pt-BR' },
-
 	],
 	bootstrap: [AppComponent],
 })
