@@ -2,35 +2,45 @@ import { MenuItems } from '../models/menu-items.model';
 
 export const menuItems: MenuItems[] = [
 	{
-    label: 'Currículo',
-    items: [
-      {
-        label: 'Início',
-        icon: 'pi pi-home',
-        routerLink: ['/'],
-      },
-      {
-        label: 'Sobre Mim',
-        icon: 'pi pi-user',
-        routerLink: ['cv', 'sobre-min'],
-      },
-      {
-        label: 'Experiências',
-        icon: 'pi pi-id-card',
-        routerLink: ['cv', 'experiencias'],
-      },
-      {
-        label: 'Formação',
-        icon: 'pi pi-home',
-        routerLink: ['cv', 'formacao'],
-      },
-      {
-        label: 'Projetos',
-        icon: 'pi pi-desktop',
-        routerLink: ['cv', 'projetos'],
-      },
-    ],
-  },
+		label: 'Currículo',
+		items: [
+			{
+				label: 'Apresentação',
+				icon: 'pi pi-home',
+				routerLink: ['/'],
+			},
+			{
+				label: 'Sobre Mim',
+				icon: 'pi pi-user',
+				routerLink: ['cv', 'sobre-min'],
+			},
+			{
+				label: 'Experiências Profissionais',
+				icon: 'pi pi-id-card',
+				routerLink: ['cv', 'experiencias'],
+			},
+			{
+				label: 'Certificações',
+				icon: 'pi pi-book',
+				routerLink: ['cv', 'certificacoes'],
+			},
+			{
+				label: 'Formações',
+				icon: 'pi pi-home',
+				routerLink: ['cv', 'formacoes'],
+			},
+			{
+				label: 'Competências e Habilidades',
+				icon: 'pi pi-list',
+				routerLink: ['cv', 'competencias'],
+			},
+			/* {
+				label: 'Projetos',
+				icon: 'pi pi-user-edit',
+				routerLink: ['cv', 'projetos'],
+			}, */
+		],
+	},
 	{
 		label: 'Contatos',
 		items: [
@@ -47,21 +57,23 @@ export const menuItems: MenuItems[] = [
 				target: '_blank',
 			},
 			{
-				label: 'Instragram',
+				label: 'Instagram',
 				icon: 'pi pi-instagram',
 				url: ['https://www.instagram.com/lucas.speixoto/'],
 				target: '_blank',
 			},
-      {
+			{
 				label: 'GitHub',
 				icon: 'pi pi-github',
 				url: ['https://github.com/lucasspeixoto'],
 				target: '_blank',
 			},
-      {
+			{
 				label: 'WhatsApp',
 				icon: 'pi pi-whatsapp',
-				url: ['https://api.whatsapp.com/send?phone=5519982621117&text=Ol%C3%A1%21+Lucas, tudo bem ?&lang=pt_pt'],
+				url: [
+					'https://api.whatsapp.com/send?phone=5519982621117&text=Ol%C3%A1%21+Lucas, tudo bem ?&lang=pt_pt',
+				],
 				target: '_blank',
 			},
 		],
