@@ -1,3 +1,4 @@
 export const environment = {
-	production: false,
+	API_URL: 'https://lucasspeixoto.github.io/profile',
+	production: true,
 };
