@@ -1,5 +1,5 @@
 module.exports = {
-	rootTranslationsPath: 'src/assets/i18n/',
+	rootTranslationsPath: 'docs/assets/i18n/',
 	langs: ['en-us', 'es-es', 'pt-br'],
 	keysManager: {},
 };
